GS-hw: Expand hw blend in more cases.
BLEND_ZERO:
Expand it to when output is 0, no need to run sw blend here.

BLEND_CLAMP:
On colclap 1 negative value will be clamped to 0 (0 - (Cd*As/Ad/Af)) or when Cd/Cs blends with Alpha Cd*(1 - As/F)
In Cd/Ad math cases it will help Vulkan and OpenGL run less barriers or hdr, for DX11 less hdr and framebuffer reading.